fixed master layout re rendering issue
